+++ CHANGELOG.md
@@ -11,7 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix Mongo driver vulnerability [#123](https://github.com/rokwire/core-building-block/issues/123)
 
 ### Added
-- Fix admin APIs after the model changes [#173] (https://github.com/rokwire/core-building-block/issues/173)
+- Set up delete account endpoint [#180](https://github.com/rokwire/core-building-block/issues/180)
+- Anonymous profile(non-pii) endpoints [#135](https://github.com/rokwire/core-building-block/issues/135)
+- User PII endpoints [#128](https://github.com/rokwire/core-building-block/issues/128)
 - Handle refresh tokens across multiple devices/apps/orgs [#149](https://github.com/rokwire/core-building-block/issues/149)
 - Expose admin API which gets applications list [#104](https://github.com/rokwire/core-building-block/issues/104)
 - Restructure auth package to provide APIs interface [#161](https://github.com/rokwire/core-building-block/issues/161)
@@ -50,6 +52,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Set up logging [#6](https://github.com/rokwire/core-building-block/issues/6)
 
 ### Fixed
+- Fix admin APIs after the model changes [#173] (https://github.com/rokwire/core-building-block/issues/173)
+- Fix login issues [#178](https://github.com/rokwire/core-building-block/issues/178)
+- Fix base path validation issue [#174](https://github.com/rokwire/core-building-block/issues/174)
 - Fix auth credentials search for multiple apps [#153](https://github.com/rokwire/core-building-block/issues/153)
 - Fix GlobalPermission and OrganizationPermission in the doc APIs model [#151](https://github.com/rokwire/core-building-block/issues/151)
 - OIDC auth bug fixes [#143](https://github.com/rokwire/core-building-block/issues/143)
@@ -57,5 +62,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixing the Log and the Changelog for issues #35 and #36 [#54](https://github.com/rokwire/core-building-block/issues/54)
 
 ### Changed
+- Move temporary claims to auth library [#183](https://github.com/rokwire/core-building-block/issues/183)
+- Prepare the service to be deployed into Rokwire infrastructure [#176](https://github.com/rokwire/core-building-block/issues/176)
 - Users authentication polish [#155](https://github.com/rokwire/core-building-block/issues/155)
 - Optimise the Mongo DB collections indexes usage [#146](https://github.com/rokwire/core-building-block/issues/146)